refactor(args): mv into a separate module
