+++ src/main.rs
@@ -5,15 +5,14 @@ use polars::lazy::frame::IntoLazy;
 use polars_sql::SQLContext;
 use ratatui::backend::CrosstermBackend;
 use ratatui::Terminal;
-use std::fmt::Display;
 use std::io;
-use std::str::FromStr;
 use tabiew::app::{AppResult, StatusBar, Table};
 use tabiew::command::CommandList;
 use tabiew::event::{Event, EventHandler};
 use tabiew::handler::handle_key_events;
 use tabiew::tui::Tui;
 use tabiew::utils::column_type_brute_foce;
+use tabiew::args::{Args, InferSchema};
 
 fn main() -> AppResult<()> {
     // Parse CLI
@@ -105,80 +104,3 @@ fn main() -> AppResult<()> {
     Ok(())
 }
 
-#[derive(Parser, Debug)]
-#[command(version, about, long_about = None)]
-struct Args {
-    #[arg(help = "File to open", required = true)]
-    file_name: String,
-
-    #[arg(
-        long,
-        help = "CSV file contains no header row",
-        default_value_t = false
-    )]
-    no_header: bool,
-
-    #[arg(
-        long,
-        help = "Ignore parsing errors while loading the CSV",
-        default_value_t = false
-    )]
-    ignore_errors: bool,
-
-    #[arg(
-        long,
-        help = "Schema inference method {no, fast, full, safe}.",
-        required = false,
-        default_value_t = InferSchema::Fast,
-    )]
-    infer_schema: InferSchema,
-
-    #[arg(
-        long,
-        help = "The field separator or delimiter",
-        required = false,
-        default_value_t = ','
-    )]
-    separator: char,
-
-    #[arg(
-        long,
-        help = "Quote character",
-        required = false,
-        default_value_t = '"'
-    )]
-    quote_char: char,
-}
-
-#[derive(Debug, Clone)]
-enum InferSchema {
-    No,
-    Fast,
-    Full,
-    Safe,
-}
-
-impl FromStr for InferSchema {
-    type Err = &'static str;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "no" => Ok(InferSchema::No),
-            "fast" => Ok(InferSchema::Fast),
-            "full" => Ok(InferSchema::Full),
-            "safe" => Ok(InferSchema::Safe),
-            _ => Err("Invalid schema inference option"),
-        }
-    }
-}
-
-impl Display for InferSchema {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            InferSchema::No => write!(f, "no"),
-            InferSchema::Fast => write!(f, "fast"),
-            InferSchema::Full => write!(f, "full"),
-            InferSchema::Safe => write!(f, "safe"),
-        }
-    }
-}
